requirements.txt
transformers
torch
sentence-transformers
faiss-cpu
langchain
pdfplumber
tqdm
protobuf
sentencepiece
requests
wikipedia
langchain-community
langchain-huggingface